Move recipe preview rendering to new module
